Moved files into initial stdmicro structure
