--- Pins.h
@@ -1,6 +1,0 @@
-#ifndef __PINS_H
-#define __PINS_H
-
-#include "types.h"
-
-#endif
