+++ e-ink.cpp
@@ -1,17 +1,16 @@
 #include "Led.h"
-#include "Pins.h"
-#include "GD7965.h"
+#include "hardware/display/eink/gd7965/GD7965.h"
 #include "Font.h"
 #include "Path.h"
 #include "DejaVuSans.h"
 #include "DejaVuSerif.h"
 #include "Frequencies.h"
 
-#include "RP2040_I2C.h"
-#include "RP2040_SPI.h"
-#include "RP2040_GPIO.h"
+#include "hardware/processor/rp2040/RP2040_I2C.h"
+#include "hardware/processor/rp2040/RP2040_SPI.h"
+#include "hardware/processor/rp2040/RP2040_GPIO.h"
 
-#include "DS3231.h"
+#include "hardware/RTC/ds3231/DS3231.h"
 
 #include "pico/stdlib.h"
 #include "hardware/gpio.h"
