+++ stdmicro/common.h
